separate pipeline schemas
